refactor:dm-card css 수정 및 height 제한 해결
