+++ app/app.css
@@ -5,7 +5,7 @@
 
 html,
 body {
-  @apply bg-white dark:bg-gray-950;
+  @apply min-h-screen bg-white dark:bg-gray-950;
 
   @media (prefers-color-scheme: dark) {
     color-scheme: dark;
