+++ app/features/ideas/components/idea-card.tsx
